+++ Telegram/Watch/Bridge/TGBridgeBotSignals.h
@@ -2,7 +2,7 @@
 
 @interface TGBridgeBotSignals : NSObject
 
-+ (SSignal *)botInfoForUserId:(int32_t)userId;
++ (SSignal *)botInfoForUserId:(int64_t)userId;
 + (SSignal *)botReplyMarkupForPeerId:(int64_t)peerId;
 
 @end
