+++ Telegram/Watch/Extension/TGNeoBubbleMessageViewModel.m
@@ -28,7 +28,7 @@ const CGFloat TGNeoBubbleHeaderSpacing = 2.0f;
         
         if (!message.outgoing && type == TGNeoMessageTypeGroup)
         {
-            _authorNameModel = [[TGNeoLabelViewModel alloc] initWithText:[users[@(message.fromUid)] displayName] font:[UIFont systemFontOfSize:14] color:[TGColor colorForUserId:(int32_t)message.fromUid myUserId:context.userId] attributes:nil];
+            _authorNameModel = [[TGNeoLabelViewModel alloc] initWithText:[users[@(message.fromUid)] displayName] font:[UIFont systemFontOfSize:14] color:[TGColor colorForUserId:(int64_t)message.fromUid myUserId:context.userId] attributes:nil];
             [self addSubmodel:_authorNameModel];
         }
