+++ Telegram/Watch/WatchCommonWatch/TGBridgeContactMediaAttachment.h
@@ -2,7 +2,7 @@
 
 @interface TGBridgeContactMediaAttachment : TGBridgeMediaAttachment
 
-@property (nonatomic, assign) int32_t uid;
+@property (nonatomic, assign) int64_t uid;
 @property (nonatomic, strong) NSString *firstName;
 @property (nonatomic, strong) NSString *lastName;
 @property (nonatomic, strong) NSString *phoneNumber;
